Undo renaming of bulk case event
