+++ src/test/java/uk/gov/hmcts/divorce/systemupdate/event/SystemUpdateCaseWithCourtHearingTest.java
@@ -29,7 +29,7 @@ import static uk.gov.hmcts.divorce.testutil.ConfigTestUtil.getEventsFrom;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.validCaseWithCourtHearing;
 
 @ExtendWith(MockitoExtension.class)
-class SystemUpdateBulkCaseWithCourtHearingTest {
+class SystemUpdateCaseWithCourtHearingTest {
 
     @Mock
     private GenerateCertificateOfEntitlement generateCertificateOfEntitlement;
